requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Django==4.1.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-environ==0.10.0
django-timezone-field==5.0
djangorestframework==3.14.0
docopt==0.6.2
fastapi==0.92.0
fernet==1.0.1
h11==0.13.0
idna==3.3
Jinja2==3.1.2
jose==1.0.0
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
outcome==1.2.0
pandas==1.5.3
pipreqs==0.4.11
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyaes==1.6.1
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
pyusb==1.2.1
qrcode==7.4.2
redis==4.5.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starlette==0.25.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.10
usb==0.0.83.dev0
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==3.8.2
wsproto==1.1.0
yarg==0.1.9